Adjusting the dialogBox elements into the resolution of the game.
Now the elements of the dialogBox like the Box itself, TextField and the
Responses are integrated with the screen. If the game is played in a
different resolution, the elements will adapt automatically to it.

Now these elements have an anchor 0:0 related to the canvas/panel qhere
they are insert into. The same was done to the Minimap, and the Arrow is
inside the image of the minimap, with an anchor 0.5:0.5 and it remains
in the middle.

The problem generated is that the minimap camera must adjust its
resolutions as well.